Remove the "Delete Preview Environment" workflow - too many permissions issues
It's not possible to delete preview environment without a privileged
access token. This is too much risk for a chore workflow. We don't want
every PR to have access to a privileged token that can perform repo admin
actions.

So just leave this as a manual chore we can do occasionally.