Use a dedicated `deploy` branch for GitHub-pages deployments
We'll use GitHub Pages for the live site, but we can only have
one GitHub pages subdomain for each repo, so we cannot use it
for preview environments for each PR. Similarly, we don't want
every push to `main` to trigger a deployment to the live website.

Instead, configure the workflow to run on push to a dedicated branch
called `deploy`, so that we can explicitly choose when we want to
deploy the site to the live web.

Also update the Readme to describe this workflow and include a
link to a pre-filled Pull Request editor for merging changes from
the `main` branch targeting the `deploy` branch.